Remove magic numbers, split code to modules
