+++ js/mock-data.js
@@ -1,0 +1,174 @@
+export default {
+  photoDescriptions: [
+    'Прекрасный закат на пляже',
+    'Удивительные горы в тумане',
+    'Романтический вечер на балконе',
+    'Цветущие вишневые деревья в парке',
+    'Заколдованный лес в осенние дни',
+    'Магический водопад в горах',
+    'Забавные животные в зоопарке',
+    'Архитектурная красота старого города',
+    'Игра света и тени в городском парке',
+    'Морской пейзаж с ярким закатом',
+    'Вдохновляющие скульптуры в саду',
+    'Уютный камин в домашней гостиной',
+    'Исторический замок на холме',
+    'Расцветающие поля под лучами солнца',
+    'Волшебный взгляд на ночной город',
+    'Летящие воздушные шары в небе',
+    'Сказочный замок в горах',
+    'Морские волны на берегу',
+    'Таинственные руины старого храма',
+    'Лесной аромат и свежий воздух в горах',
+    'Мир сквозь объектив фотокамеры',
+    'Яркие цветы в ботаническом саду',
+    'Солнечный день на пляже',
+    'Семейный пикник в парке',
+    'Индустриальный пейзаж в городе',
+    'Космический вид на город из окна самолета',
+    'Игра цвета и текстуры в городской архитектуре',
+    'Мистический вид на затерянный город',
+    'Дикие животные в их естественной среде обитания',
+    'Экзотические фрукты на рыночной ярмарке',
+    'Золотистые поля в закатном свете',
+    'Живописный вид на городской порт',
+    'Динамичная городская жизнь в черно-белом исполнении',
+    'Потрясающий вид на горные вершины',
+    'Классическая архитектура итальянского города',
+    'Сказочный вид на замерзшее озеро',
+    'Таинственная дорога в густом лесу',
+    'Роскошный вид с вершины небоскреба',
+    'Вдохновляющий закат в горах',
+    'Игра света на стеклянных фасадах зданий',
+    'Природная красота национального парка',
+    'Абстрактные узоры на песке пустыни',
+    'Магический вид на водопад с высоты птичьего полета',
+    'Променад по набережной вечернего города',
+    'Интересные граффити на стенах городских зданий',
+    'Умиротворенная атмосфера в ботаническом саду',
+    'Загадочные фигуры в стиле стимпанк',
+    'Закатная атмосфера на поляне в лесу',
+    'Яркие огни ночного города',
+    'Таинственный вид на старинный маяк',
+    'Сказочный вид на горные вершины в сумеречном свете',
+    'Умиротворенная атмосфера в японском саду',
+    'Загадочное здание в стиле стимпанк',
+    'Потрясающий вид на горные вершины в сумеречном свете',
+    'Вдохновляющий закат в горах',
+    'Игра света на стеклянных фасадах зданий',
+    'Природная красота национального парка',
+    'Абстрактные узоры на песке пустыни',
+    'Магический вид на водопад с высоты птичьего полета',
+    'Променад по набережной вечернего города',
+    'Интересные граффити на стенах городских зданий',
+    'Умиротворенная атмосфера в ботаническом саду',
+    'Загадочные фигуры в стиле стимпанк',
+    'Закатная атмосфера на поляне в лесу',
+    'Яркие огни ночного города',
+    'Таинственный вид на старинный маяк',
+    'Сказочный вид на горные вершины в сумеречном свете',
+    'Умиротворенная атмосфера в японском саду',
+    'Загадочное здание в стиле стимпанк',
+    'Потрясающий вид на горные вершины в сумеречном свете',
+  ],
+  photoComments: [
+    'Прекрасная композиция и выбор цветовой гаммы делают этот снимок действительно интересным',
+    'Идеальный момент, запечатленный на этом снимке, это действительно великолепно',
+    'Превосходная атмосфера на этом снимке, это действительно восхитительно',
+    'Невероятное внимание к деталям делает этот снимок просто безупречным',
+    'Использование контраста и глубины на этом снимке просто восхитительно',
+    'Использование цвета на этом снимке просто великолепно, это действительно вдохновляет',
+    'Это восхитительное произведение искусства, каждая мелочь продумана до мелочей',
+    'Очень вдохновляющий снимок, который заставляет задуматься об удивительной природе',
+    'Этот снимок просто восхитителен, атмосфера на нем просто захватывает воображение',
+    'Исключительный контраст и композиция делают этот снимок действительно уникальным',
+    'Удивительное использование света и тени делает этот снимок действительно превосходным',
+    'Прекрасная композиция и выбор цветовой гаммы делают этот снимок действительно интересным',
+    'Идеальный момент, запечатленный на этом снимке, это действительно великолепно',
+    'Превосходная атмосфера на этом снимке, это действительно восхитительно',
+    'Невероятное внимание к деталям делает этот снимок просто безупречным',
+    'Использование контраста и глубины на этом снимке просто восхитительно',
+    'Использование цвета на этом снимке просто великолепно, это действительно вдохновляет',
+    'Это восхитительное произведение искусства, каждая мелочь продумана до мелочей',
+    'Очень вдохновляющий снимок, который заставляет задуматься об удивительной природе',
+    'Этот снимок просто восхитителен, атмосфера на нем просто захватывает воображение',
+    'Фантастический снимок, каждая деталь на нем просто потрясающая',
+    'Отличная работа с освещением, это действительно добавляет эмоционального оттенка снимку',
+    'Этот снимок просто захватывает воображение, это действительно потрясающе',
+    'Исключительное использование цветовых оттенков делает этот снимок действительно великолепным',
+    'Этот снимок просто великолепен, атмосфера на нем действительно впечатляет',
+    'Прекрасная композиция и выбор ракурса делают этот снимок действительно интересным',
+    'Идеально подобранный момент для съемки, это действительно великолепно',
+    'Превосходный выбор темы для снимка, это действительно восхитительно',
+    'Удивительное внимание к мелочам делает этот снимок просто потрясающим',
+    'Использование глубины резкости и перспективы делают этот снимок действительно уникальным',
+    'Удивительное использование света и тени делает этот снимок действительно превосходным',
+    'Прекрасная композиция и выбор цветовой гаммы делают этот снимок действительно интересным',
+    'Идеальный момент, запечатленный на этом снимке, это действительно великолепно',
+    'Превосходная атмосфера на этом снимке, это действительно восхитительно',
+    'Невероятное внимание к деталям делает этот снимок просто безупречным',
+    'Использование контраста и глубины на этом снимке просто восхитительно',
+    'Использование цвета на этом снимке просто великолепно, это действительно вдохновляет',
+    'Это восхитительное произведение искусства, каждая мелочь продумана до мелочей',
+    'Очень вдохновляющий снимок, который заставляет задуматься об удивительной природе',
+    'Этот снимок просто восхитителен, атмосфера на нем просто захватывает воображение',
+    'Великолепный снимок, каждая деталь на нем просто поражает воображение',
+    'Отличная игра света и тени, это действительно придает глубину снимку',
+    'Этот снимок просто захватывает дух, это действительно восхитительно',
+    'Исключительное использование цветовых оттенков делает этот снимок действительно превосходным',
+    'Этот снимок просто восхитителен, атмосфера на нем действительно поражает',
+    'Прекрасная композиция и выбор ракурса делают этот снимок действительно интересным',
+    'Идеально подобранный момент для съемки, это действительно великолепно',
+    'Превосходный выбор темы для снимка, это действительно восхитительно',
+    'Удивительное внимание к деталям делает этот снимок просто поразительным',
+    'Использование глубины резкости и перспективы делают этот снимок действительно уникальным',
+    'Удивительное использование света и тени делает этот снимок действительно превосходным',
+    'Прекрасная композиция и выбор цветовой гаммы делают этот снимок действительно интересным',
+    'Идеальный момент, запечатленный на этом снимке, это действительно великолепно',
+    'Превосходная атмосфера на этом снимке, это действительно восхитительно',
+    'Невероятное внимание к деталям делает этот снимок просто безупречным',
+    'Использование контраста и глубины на этом снимке просто восхитительно',
+    'Использование цвета на этом снимке просто великолепно, это действительно вдохновляет',
+    'Это восхитительное произведение искусства, каждая мелочь продумана до мелочей',
+    'Очень вдохновляющий снимок, который заставляет задуматься об удивительной природе',
+    'Этот снимок просто восхитителен, атмосфера на нем просто захватывает воображение',
+  ],
+  userNames: [
+    'Алексей',
+    'Анна',
+    'Владимир',
+    'Вера',
+    'Георгий',
+    'Дмитрий',
+    'Екатерина',
+    'Иван',
+    'Ирина',
+    'Кирилл',
+    'Людмила',
+    'Мария',
+    'Наталья',
+    'Олег',
+    'Петр',
+    'Сергей',
+    'София',
+    'Татьяна',
+    'Ульяна',
+    'Федор',
+    'Харитон',
+    'Царица',
+    'Шах',
+    'Эмилия',
+    'Юлия',
+    'Ярослав',
+  ],
+
+  getPhotoDescriptions: function () {
+    return this.photoDescriptions;
+  },
+  getPhotoComments: function () {
+    return this.photoComments;
+  },
+  getUserNames: function () {
+    return this.userNames;
+  },
+};
